+++ ts-sdk/actions/src/increaseLiquidity.ts
@@ -1,0 +1,67 @@
+import { Address, Lamports } from "@solana/web3.js";
+import {
+  IncreaseLiquidityQuoteParam,
+  openPositionInstructions,
+  openFullRangePositionInstructions,
+  increaseLiquidityInstructions,
+} from "@orca-so/whirlpools";
+import { IncreaseLiquidityQuote } from "@orca-so/whirlpools-core";
+import { executeWhirlpoolInstruction } from "./helpers";
+
+// Open a concentrated liquidity position
+export async function openConcentratedPosition(
+  poolAddress: Address,
+  tokenAmount: IncreaseLiquidityQuoteParam,
+  lowerPrice: number,
+  upperPrice: number,
+  slippageToleranceBps?: number
+): Promise<{
+  callback: () => Promise<string>;
+  quote: IncreaseLiquidityQuote;
+  initializationCost: Lamports;
+  positionMint: Address;
+}> {
+  return executeWhirlpoolInstruction(
+    openPositionInstructions,
+    poolAddress,
+    tokenAmount,
+    lowerPrice,
+    upperPrice,
+    slippageToleranceBps
+  );
+}
+
+// Open a full range position
+export async function openFullRangePosition(
+  poolAddress: Address,
+  tokenAmount: IncreaseLiquidityQuoteParam,
+  slippageToleranceBps?: number
+): Promise<{
+  callback: () => Promise<string>;
+  quote: IncreaseLiquidityQuote;
+  initializationCost: Lamports;
+}> {
+  return executeWhirlpoolInstruction(
+    openFullRangePositionInstructions,
+    poolAddress,
+    tokenAmount,
+    slippageToleranceBps
+  );
+}
+
+// Increase liquidity in an existing position
+export async function increasePosLiquidity(
+  positionMintAddress: Address,
+  tokenAmount: IncreaseLiquidityQuoteParam,
+  slippageToleranceBps?: number
+): Promise<{
+  callback: () => Promise<string>;
+  quote: IncreaseLiquidityQuote;
+}> {
+  return executeWhirlpoolInstruction(
+    increaseLiquidityInstructions,
+    positionMintAddress,
+    tokenAmount,
+    slippageToleranceBps
+  );
+}
