+++ ts-sdk/actions/src/index.ts
@@ -1,5 +1,8 @@
-export * from "./pools";
-export * from "./positions";
-export * from "./swap";
-export * from "./harvestFees";
 export * from "./config";
+export * from "./createPool";
+export * from "./decreaseLiquidity";
+export * from "./harvest";
+export * from "./increaseLiquidity";
+export * from "./pool";
+export * from "./position";
+export * from "./swap";
