+++ ts-sdk/actions/src/position.ts
@@ -1,0 +1,4 @@
+export {
+  fetchPositionsForOwner,
+  fetchPositionsInWhirlpool,
+} from "@orca-so/whirlpools";
